move pie chart logic into own component, remove unused code
